+++ lib/pages/dashboards/dashboards_carousel_page.dart
@@ -4,6 +4,7 @@ import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/pages/dashboards/dashboard_page.dart';
+import 'package:lotti/pages/dashboards/how_to_use_page.dart';
 import 'package:lotti/utils/sort.dart';
 import 'package:lotti/widgets/app_bar/title_app_bar.dart';
 import 'package:lotti/widgets/charts/empty_dashboards_widget.dart';
@@ -35,7 +36,7 @@ class _DashboardCarouselPageState extends State<DashboardCarouselPage> {
         final dashboards = filteredSortedDashboards(snapshot.data ?? []);
 
         if (dashboards.isEmpty) {
-          return const EmptyDashboards();
+          return const HowToUsePage();
         }
 
         final rangeStart = getRangeStart(context: context);