+++ lib/pages/dashboards/how_to_use_page.dart
@@ -1,0 +1,68 @@
+import 'package:auto_size_text/auto_size_text.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:lotti/themes/theme.dart';
+import 'package:lotti/utils/platform.dart';
+import 'package:url_launcher/url_launcher.dart';
+
+class HowToUsePage extends StatefulWidget {
+  const HowToUsePage({
+    super.key,
+  });
+
+  @override
+  State<HowToUsePage> createState() => _HowToUsePageState();
+}
+
+class _HowToUsePageState extends State<HowToUsePage> {
+  bool isHovering = false;
+
+  @override
+  Widget build(BuildContext context) {
+    final localizations = AppLocalizations.of(context)!;
+
+    return Scaffold(
+      backgroundColor: colorConfig().negspace,
+      //appBar: const DashboardsAppBar(),
+      body: Stack(
+        children: [
+          Align(
+            child: OutlinedButton(
+              onHover: (hovering) => setState(() {
+                isHovering = hovering;
+              }),
+              onPressed: () {
+                final uri = Uri.parse(
+                  'https://github.com/matthiasn/lotti/blob/main/docs/MANUAL.md',
+                );
+                launchUrl(uri);
+              },
+              style: ElevatedButton.styleFrom(
+                backgroundColor:
+                    isHovering ? colorConfig().riptide : colorConfig().negspace,
+                side: const BorderSide(),
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(100),
+                ),
+                padding: EdgeInsets.symmetric(
+                  vertical: isMobile ? 20 : 30,
+                  horizontal: isMobile ? 30 : 45,
+                ),
+              ),
+              child: AutoSizeText(
+                localizations.dashboardsHowToHint,
+                style: TextStyle(
+                  color: colorConfig().coal,
+                  fontSize: 25,
+                  fontFamily: 'PlusJakartaSans',
+                  fontWeight: FontWeight.w400,
+                ),
+                maxLines: 2,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}